+++ src/Script/gameObjs/oth/PlayerGoCircle.cpp
@@ -63,7 +63,7 @@ void PlayerGoCircle::init(GameObj &goRef_,const DyParam &dyParams_ ){
                                 "idle", 
                                 RenderLayerType::MajorGoes, //- 不设置 固定zOff值  
                                 &esrc::get_shaderRef(ShaderType::PlayerGoCircle),  // pic shader
-                                glm::vec2{ 0.0f, 0.0f }, //- pposoff
+                                glm::dvec2{ 0.0, 0.0 }, //- pposoff
                                 -500.0,  //- off_z， 沉在所有 MajorGo 后方
                                 true //- isVisible
                                 );
@@ -75,7 +75,7 @@ void PlayerGoCircle::init(GameObj &goRef_,const DyParam &dyParams_ ){
                                 "idle",
                                 RenderLayerType::MajorGoes, //- 不设置 固定zOff值  
                                 &esrc::get_shaderRef(ShaderType::PlayerGoCircle),  // pic shader
-                                glm::vec2{ 0.0f, 0.0f }, //- pposoff
+                                glm::dvec2{ 0.0, 0.0 }, //- pposoff
                                 500.0,  //- off_z， 浮在所有 MajorGo 前方
                                 true //- isVisible
                                 );
